+++ netpyne/utils.py
@@ -322,7 +322,10 @@ def getCellParams(cell, varList, origGlob):
                         varNameSplit = varName
                         if varName not in ignoreVars:
                             try:
-                                varVals = [seg.__getattribute__(varNameSplit+ionName) for seg in sec]
+                                if varNameSplit in ['i','o']: # var name after ion name (eg. 'nai', 'nao')
+                                    varVals = [seg.__getattribute__(ionName+varNameSplit) for seg in sec]
+                                else: # var name before ion name (eg. 'ena')
+                                    varVals = [seg.__getattribute__(varNameSplit+ionName) for seg in sec]
                                 if len(set(varVals)) == 1:
                                     varVals = varVals[0] 
                                 ionDic[ionName][varNameSplit] = varVals